renaming project from ab_testing_tool_app -> ab_tool
