+++ ab_testing_tool/urls.py
@@ -3,5 +3,5 @@ from django.contrib import admin
 admin.autodiscover()
 
 urlpatterns = patterns('',
-    url(r'^lti_tools/ab_tool/', include("ab_testing_tool_app.urls", namespace="ab", app_name="ab_tool")),
+    url(r'^lti_tools/ab_tool/', include("ab_tool.urls", namespace="ab", app_name="ab_tool")),
 )
