+++ ab_tool/templates/ab_tool/edit_intervention_point.html
