+++ ab_tool/templates/ab_tool/modules_page_edit_intervention_point.html
