+++ ab_tool/tests/common.py
@@ -4,7 +4,7 @@ from django.utils.importlib import import_module
 from json import dumps
 from mock import patch, MagicMock
 
-from ab_testing_tool_app.constants import ADMINS
+from ab_tool.constants import ADMINS
 from error_middleware.exceptions import DEFAULT_ERROR_STATUS, RenderableError
 from error_middleware.middleware import ERROR_TEMPLATE
