+++ ab_tool/urls.py
@@ -1,15 +1,15 @@
 from django.conf.urls import patterns, url, include
 from django.contrib import admin
 
-from ab_testing_tool_app.views.main_pages import (render_intervention_point_control_panel, not_authorized,
+from ab_tool.views.main_pages import (render_intervention_point_control_panel, not_authorized,
      tool_config, download_data)
 
-from ab_testing_tool_app.views.selection_pages import (resource_selection, submit_selection,
+from ab_tool.views.selection_pages import (resource_selection, submit_selection,
                     submit_selection_new_intervention_point)
-from ab_testing_tool_app.views.intervention_point_pages import (create_intervention_point, submit_create_intervention_point,
+from ab_tool.views.intervention_point_pages import (create_intervention_point, submit_create_intervention_point,
                     edit_intervention_point,  delete_intervention_point, deploy_intervention_point, submit_edit_intervention_point,
                     modules_page_edit_intervention_point)
-from ab_testing_tool_app.views.track_pages import (create_track, submit_create_track,
+from ab_tool.views.track_pages import (create_track, submit_create_track,
                     submit_edit_track, edit_track, delete_track, finalize_tracks)
 
 admin.autodiscover()
