+++ ab_tool/views/main_pages.py
@@ -8,12 +8,12 @@ from django.template.defaultfilters import slugify
 from django.template import loader
 from ims_lti_py.tool_config import ToolConfig
 
-from ab_testing_tool_app.canvas import get_lti_param
-from ab_testing_tool_app.controllers import (get_uninstalled_intervention_points,
+from ab_tool.canvas import get_lti_param
+from ab_tool.controllers import (get_uninstalled_intervention_points,
     get_modules_with_items, get_incomplete_intervention_points)
-from ab_testing_tool_app.models import (InterventionPoint, Track, CourseStudent,
+from ab_tool.models import (InterventionPoint, Track, CourseStudent,
     CourseSettings)
-from ab_testing_tool_app.constants import ADMINS
+from ab_tool.constants import ADMINS
 
 
 def not_authorized(request):
