+++ ab_tool/views/track_pages.py
@@ -1,13 +1,13 @@
 from django.shortcuts import render_to_response, redirect, get_object_or_404
 from django_auth_lti.decorators import lti_role_required
 
-from ab_testing_tool_app.constants import ADMINS
-from ab_testing_tool_app.models import Track, CourseSettings, InterventionPoint
-from ab_testing_tool_app.canvas import get_lti_param
-from ab_testing_tool_app.exceptions import (UNAUTHORIZED_ACCESS,
+from ab_tool.constants import ADMINS
+from ab_tool.models import Track, CourseSettings, InterventionPoint
+from ab_tool.canvas import get_lti_param
+from ab_tool.exceptions import (UNAUTHORIZED_ACCESS,
     COURSE_TRACKS_ALREADY_FINALIZED, NO_TRACKS_FOR_COURSE)
 from django.http.response import HttpResponse
-from ab_testing_tool_app.controllers import (post_param, get_incomplete_intervention_points)
+from ab_tool.controllers import (post_param, get_incomplete_intervention_points)
 
 
 @lti_role_required(ADMINS)
